refactor the major methods from orm to milvus client (#70)
In pymilvus SDK, the
[ORM](https://milvus.io/api-reference/pymilvus/v2.5.x/ORM/Collection/Collection.md),
and [milvus
client](https://milvus.io/api-reference/pymilvus/v2.5.x/MilvusClient/Client/MilvusClient.md)
are two different layer/approach SDKs.
In the future, we recommend using milvus client SDK to get access and
operate Mivus service instead of ORM SDK, so we have to refactor the
usage of ORM in langchain-milvus. This PR changed the major CRUD public
methods in langchain-milvus class.
BTW, the SearchResult will be removed in the future version of pymilvus,
and it will be replaced by the List[List[dict]]

---------

Signed-off-by: ChengZi <chen.zhang@zilliz.com>